Implement star rating system and gallery view improvements
- Add star rating system with 1-5 stars for satisfaction tracking
- Create category-based ticket organization with 7 categories
- Implement gallery view with grid/list toggle and statistics
- Add Core Data model updates for rating and category fields
- Create unified card sizing for consistent gallery display
- Add margins to gallery cards for better content spacing
- Align rating field styling in ticket creation form

+++ TicketScan/TicketScan/Views/TicketShareView.swift
@@ -46,7 +46,7 @@ struct TicketShareView: View {
                         HStack {
                             Image(systemName: "calendar")
                                 .foregroundColor(.blue)
-                            Text(ticket.eventDate.formatted(date: .long, time: .shortened))
+                            Text(ticket.eventDate.formatted(date: .numeric, time: .shortened))
                                 .font(.headline)
                         }
                         
@@ -108,7 +108,7 @@ struct TicketShareView: View {
                     
                     HStack {
                         Image(systemName: "calendar")
-                        Text(ticket.eventDate.formatted(date: .long, time: .shortened))
+                        Text(ticket.eventDate.formatted(date: .numeric, time: .shortened))
                             .font(.headline)
                     }
